better AI handling of emergency stop procedure, simplified lod visibility calculations, adjustable track geometry fidelity

+++ Model3d.cpp
@@ -28,7 +28,7 @@ Copyright (C) 2001-2004  Marcin Wozniak, Maciej Czapkiewicz and others
 
 using namespace Mtable;
 
-double TSubModel::fSquareDist = 0;
+float TSubModel::fSquareDist = 0.f;
 size_t TSubModel::iInstance; // numer renderowanego egzemplarza obiektu
 texture_handle const *TSubModel::ReplacableSkinId = NULL;
 int TSubModel::iAlpha = 0x30300030; // maska do testowania flag tekstur wymiennych
